refactoring: no production code , creaded files alertFunctions,CleanUp from which we exporting things to main app, make app more readable
