+++ src/client/index.js
@@ -1,5 +1,9 @@
 import './sass/main.scss'
 
 import './js/app'
-import './js/handleNav'
-import './js/getTime'
+import {
+    handleNav
+} from './js/handleNav'
+import {
+    getTime
+} from './js/getTime'
