+++ src/client/js/alertFunctions.js
@@ -1,0 +1,11 @@
+// alert NO destination
+export const alertFn = () => {
+    alert("😊 Please, enter your a travel destination ✈️ and the start date for travel 📅");
+}
+
+
+// alert with days
+
+export const alertMoreDays = () => {
+    alert("🗓️ Sorry, but this app only covers weather 16 days in advance.\n Please enter a valid date. 🙈");
+}
