+++ src/client/js/cleanUp.js
@@ -1,13 +1,22 @@
-// import './app'
+import {
+    btnDelete,
+    counddownTitle,
+    timeCards,
+    temp,
+    enterCity,
+    weatherDescription,
+    imgCountry
+} from './app'
 
-// //Delete info about trip 
-// export const cleanUp = () => {
-//     temp.innerHTML = "";
-//     weatherDescription.innerHTML = "";
-//     imgCountry.src = "";
-//     enterCity.innerHTML = "";
-//     counddownTitle.classList.remove('active');
-//     timeCards.classList.remove('active');
-//     btnDelete.classList.remove('active')
-//     imgCountry.classList.remove('active')
-// }
+
+//Delete info about trip 
+export const cleanUp = () => {
+    temp.innerHTML = "";
+    weatherDescription.innerHTML = "";
+    imgCountry.src = "";
+    enterCity.innerHTML = "";
+    counddownTitle.classList.remove('active');
+    timeCards.classList.remove('active');
+    btnDelete.classList.remove('active')
+    imgCountry.classList.remove('active')
+}
